Oppdate bouncycastle to jdk18on and remove use of deprecated apis
The jdk15 build of bouncy castle has vulerabilities and is no longer
maintained. I upgrade to jdk18on and fix errors because of removal
of deprecated methods with regards to fetching url from a certificate.
I opted to use the new way to get the location of ocsp instead of
the getBaseObject-solution I see others have done.
This way we hopefully get a more stable code than if we use the
getBaseObject-method which has this comment in the source:
> Needed for open types, until we have better type-guided parsing support.